chore: move src/collect-ubuntu-browser-dependencies inside src/browser-installer/ubuntu-packages/
