+++ package.json
@@ -9,11 +9,11 @@
   ],
   "scripts": {
     "build": "tsc --build && npm run copy-static && npm run build-bundles",
-    "copy-static": "copyfiles 'src/browser/client-scripts/*' 'src/**/*.json' build",
+    "copy-static": "copyfiles 'src/browser/client-scripts/*' 'src/**/[!cache]*/autogenerated/**/*.json' build",
     "build-node-bundle": "esbuild ./src/bundle/cjs/index.ts --outdir=./build/src/bundle/cjs --bundle --format=cjs --platform=node --target=ES2021",
     "build-browser-bundle": "node ./src/browser/client-scripts/build.js",
     "build-bundles": "concurrently -c 'auto' 'npm:build-browser-bundle' 'npm:build-node-bundle --minify'",
-    "resolve-ubuntu-dependencies": "ts-node ./src/collect-ubuntu-browser-dependencies",
+    "resolve-ubuntu-dependencies": "ts-node ./src/browser-installer/ubuntu-packages/collect-dependencies",
     "check-types": "tsc --project tsconfig.spec.json",
     "clean": "rimraf build/ *.tsbuildinfo",
     "lint": "eslint --cache . && prettier --check .",
