+++ src/browser-installer/ubuntu-packages/collect-dependencies/constants.ts
