+++ src/browser-installer/ubuntu-packages/collect-dependencies/shared-object.ts
