+++ src/browser-installer/ubuntu-packages/collect-dependencies/ubuntu/index.ts
