+++ src/browser-installer/ubuntu-packages/collect-dependencies/ubuntu/utils.ts
