+++ src/browser-installer/ubuntu-packages/collect-dependencies/utils.ts
