+++ tsconfig.json
@@ -6,7 +6,7 @@
         "src/**/*.test.ts",
         "src/browser/client-scripts",
         "src/bundle/cjs",
-        "src/collect-ubuntu-browser-dependencies",
+        "src/browser-installer/ubuntu-packages/collect-dependencies",
         "src/runner/browser-env/vite/browser-modules"
     ],
     "compilerOptions": {
